requirements.txt
absl-py
pytest
sacrebleu==1.4.10
six
wheel
progressbar
sklearn
future
dataclasses
jsonlines
numpy==1.22.4
tqdm
transformers==4.2.1
sentencepiece==0.1.97
tokenizers==0.9.4
rouge==1.0.1
Pillow==9.0.1
pyyaml
nltk
jupyter
notebook
pandas